requirements/base.txt
Django==1.9.1
psycopg2==2.5.4
django-leaflet
gunicorn==19.3.0
requests[security]
xlrd
djangorestframework==3.3.2
djangorestframework-csv==1.4.1
django-rest-swagger==0.3.5
django-cors-headers==1.1.0
openpyxl==2.3.3
numpy==1.10.4
pandas==0.17.1
scipy==0.17.0
google-api-python-client==1.5.1
